refactor: extract PostTranscriptionPipelineTypes.swift from TranscriptionSessionBuilder.swift
Extracts 2 pipeline result enums (PostTranscriptionPipelineMode,
PostTranscriptionPipelineResult) from TranscriptionSessionBuilder.swift into
a new PostTranscriptionPipelineTypes.swift.

Byte-identical move; SHA verified against origin/main.

Reduces TranscriptionSessionBuilder.swift from 119 to 96 lines.

Tests: 667 (unchanged).
